move type declarations
