+++ src/plugins/image/index.ts
@@ -2,7 +2,7 @@ import { CN_ACTIVE, TN_IMG } from '../../contants'
 import { exsied } from '../../core'
 import { DomUtils } from '../../core/dom_utils'
 import { addEleClickCallbackByClass, addEleClickCallbackByTag } from '../../core/events'
-import { Commands, ExsiedPlugin } from '../../types'
+import { Commands, ExsiedPlugin } from '../../core/plugin'
 import { Toolbar } from '../../ui/toolbar'
 import { CN_BTN_SETTING, CN_ICON, PLUGIN_CONF, PLUGIN_NAME, RESIZER_ID } from './base'
 import { insertImage, onClickImage, onClickImageSettingButton } from './event_handlers'
