+++ src/plugins/link/index.ts
@@ -2,7 +2,7 @@ import { CN_ACTIVE, TN_A } from '../../contants'
 import { exsied } from '../../core'
 import { DomUtils } from '../../core/dom_utils'
 import { addEleClickCallbackByTag } from '../../core/events'
-import { Commands, ExsiedPlugin } from '../../types'
+import { Commands, ExsiedPlugin } from '../../core/plugin'
 import { Toolbar } from '../../ui/toolbar'
 import { CN_ICON, PLUGIN_CONF, PLUGIN_NAME, POPUP_ID } from './base'
 import { insertLink, onClickLink } from './event_handlers'
