+++ src/ui/toolbar/index.ts
@@ -2,9 +2,9 @@ import { DATA_ATTR_CN_ICON, TN_DIV } from '../../contants'
 import { exsied } from '../../core'
 import { DomUtils } from '../../core/dom_utils'
 import { execEleEventClickCallbackByTag } from '../../core/events'
-import { PLUGINS } from '../../core/plugin'
+import { ChangeEventHandler, ClickEventHandler, PLUGINS } from '../../core/plugin'
 import { SelectionUtils } from '../../core/selection_utils'
-import { ChangeEventHandler, ClickEventHandler, HTMLTagNames } from '../../types'
+import { HTMLTagNames } from '../../types'
 import { tagNameLc } from '../../utils'
 import { uniqueArray } from '../../utils/array'
 import { CN_DDROPDOWN_LIST_SHOW, DropdownMenu, genDropdownId } from '../dropdown'
