+++ src/ui/toolbar/styles.scss
@@ -14,6 +14,7 @@
 
 .exsied-toolbar {
 	&.exsied-toolbar-main {
+		overflow: scroll;
 		position: sticky;
 		top: 0;
 	}
